+++ R/fst_hudson_pairwise.R
@@ -16,6 +16,8 @@
 #' @param mem_lim Memory limit in GB, used to break up genotype data into chunks for very large datasets.
 #' Note memory usage is somewhat underestimated and is not controlled strictly.
 #' Default in Linux and Windows is `mem_factor` times the free system memory, otherwise it is 1GB (OSX and other systems).
+#' @param m_chunk_max Sets the maximum number of loci to process at the time.
+#' Actual number of loci loaded may be lower if memory is limiting.
 #'
 #' @return A symmetric matrix of FST estimates between every pair of subpopulations.
 #' The diagonal has zero values.
@@ -62,7 +64,8 @@ fst_hudson_pairwise <- function(
                                 m = NA,
                                 loci_on_cols = FALSE,
                                 mem_factor = 0.7,
-                                mem_lim = NA
+                                mem_lim = NA,
+                                m_chunk_max = 1000
                                 ) {
     if (missing(X))
         stop('Genotype matrix `X` is required!')
@@ -120,8 +123,11 @@ fst_hudson_pairwise <- function(
                          mem_factor = mem_factor
                      )
     m_chunk <- data$m_chunk
+    # cap value to a nice performing value (very good speed, minimal memory)
+    if ( m_chunk > m_chunk_max )
+        m_chunk <- m_chunk_max
 
-        # navigate chunks
+    # navigate chunks
     i_chunk <- 1 # start of first chunk (needed for matrix inputs only; as opposed to function inputs)
     while(TRUE) { # start an infinite loop, break inside as needed
         # indexes to extract loci, and also so save to FstTs and FstBs vectors